Fixed label sim kitchen dataset to original version
